Seperate controller tests into relevant folders
